add test plan for virtio 1.0.
Virtio 1.0 is a new version of virtio. And the virtio 1.0 spec link is at
http://docs.oasis-open.org/virtio/virtio/v1.0/virtio-v1.0.pdf.

The major difference is at PCI layout. For testing virtio 1.0 pmd, we need
test the basic RX/TX, different path(txqflags), mergeable on/off, and also
test with virtio0.95 to ensure they can co-exist.

Besides, we need test virtio 1.0's performance to ensure it has similar
performance as virtio0.95.

Signed-off-by: Qian Xu <qian.q.xu@intel.com>