tests: pmdpcap - modifying the right config file + importing utils after scapy
This test used to modify config/common_linuxapp or config/common_bsdapp
to compile DPDK with PCAP, however, this config may lie as well in
config/common_base. If the PCAP variable is not found in OS specific
files, then it is assumed that config/common_base should be modified.
Scapy may contain utils itself, therefore importing dts utils is moved
after importing from scapy (so as to override utils).

Signed-off-by: Angela Czubak <aczubak at caviumnetworks.com>