tests: short_live - taking into account cavium NICs + fixes
Testpmd must be run with hardware VLAN filter disabled as thunder
driver does not support this feature - --disable-hw-vlan-filter is
appended if NIC is Cavium device.
Additional fixes done: waiting for DPDK application to be finished by
calling `ps` in a loop and checking whether the process is still alive;
sniffing in background scapy process fails at times, using tcpdump
instead; if start_up_time fails it is checked if the problem was that
time command seem not to be found (informing user about it).

Signed-off-by: Angela Czubak <aczubak at caviumnetworks.com>