tests: unit_tests_dump - taking into account cavium NICs
Testpmd must be run with hardware VLAN filter disabled as thunder
driver does not support this feature - --disable-hw-vlan-filter is
appended if NIC is Cavium device.
In test_ring_dump, matching the mbuf_pool is done against the first
element of group (instead of the last one), since it is the first group
matched.

Signed-off-by: Angela Czubak <aczubak at caviumnetworks.com>